% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adae}
\alias{adae}
\title{ADAE}
\format{
Data frame with 1191 features and 55 fields
}
\source{
\url{https://github.com/phuse-org/aesummaries/tree/main/inst/extdata/adae.xpt},
downloaded 2023-03-17
}
\usage{
adae
}
\description{
Adverse Events Analysis Dataset.
}
\keyword{datasets}
