% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_r001.R
\name{adae_r001}
\alias{adae_r001}
\title{Adverse Events Summary Table}
\usage{
adae_r001(
  datain,
  population = "Overall",
  AE_Filter = "",
  riskyn = "N",
  summary_by = "Patients",
  ctrlgrp,
  trtgrp,
  ui_lt = "AEDECOD",
  ui_ht = "AEBODSYS",
  ui_statistics = "Risk Difference",
  ui_trttotalyn = "N",
  ui_pctdisp = "TRT",
  ui_alpha = 0.05,
  ui_cutoff = 5,
  ui_sortopt = "Ascending",
  ui_sortvar = "Count"
)
}
\arguments{
\item{datain}{List containing input datasets, from \code{ae_pre_processor()}}

\item{population}{Population filter applied, for the footnote.}

\item{AE_Filter}{Adverse Event Filter applied, for the footnote}

\item{riskyn}{To calculate and display Risk Statistic or not. ("Y"/"N")}

\item{summary_by}{Measure to construct the summary by. Values: "Patients", "Events"}

\item{ctrlgrp}{Treatment Variable value which becomes control group}

\item{trtgrp}{Treatment Variable value which becomes treatment group}

\item{ui_lt}{Low Level Adverse Event term variable, used for analysis}

\item{ui_ht}{High Level Adverse Event term variable, used for analysis}

\item{ui_statistics}{Measure to be calculated. Values: "Risk Difference", "Risk Ratio"}

\item{ui_trttotalyn}{To display Total Column for Treatments or not. Values: "Y"/"N"}

\item{ui_pctdisp}{Method to set denominator for percentage incidence.
Values: "TRT", "VAR", "HT"}

\item{ui_alpha}{Alpha value to determine confidence interval for risk calculation}

\item{ui_cutoff}{Incidence Cutoff Value - subset only Higher Level terms with incidence
percentage > ui_cutoff value. All corresponding Low terms are taken.}

\item{ui_sortopt}{How to sort terms in output. Values: "Ascending", "Descending", "Alphabetical"}

\item{ui_sortvar}{Metric to sort terms by. Values: "Count", "Percent", "RiskValue"}
}
\value{
a list containing 4 objects
\itemize{
\item tout - Flextable output for display
\item rpt_data - Final data.frame used to create \code{tout}
\item title - Report title
\item footnote - Report Footnote
}
}
\description{
Adverse Events Summary Table
}
\details{
\itemize{
\item Object passed to \code{datain} must be the output of \code{ae_pre_processor},
containing elements \code{dsin}, \code{dout} and \code{bigN}. Element \code{bigN} determines whether the
output table contains (N = count) appended to each Treatment Column. Elements
\code{dsin} and \code{dout} are used to calculate counts and percentages respectively.
\item Arguments \code{population} and \code{AE_Filter} are used only in report title and footnote.
They should be the same values passed previously in the \code{ae_pre_processor()} function.
\item If \code{summary_by} is "Patients", then the count used is the unique number of participants
per term; and if it is "Events", the count used is the total number of events observed per term.
\item Arguments \code{ui_lt} and \code{ui_ht} contain the lower and higher level Event variables.
The lower level variable will be grouped by higher level, so apt variables must be passed.
Examples (Low, High): AEDECOD/AEBODSYS, FMQ_NAM/AEBODSYS or AEDECOD/FMQ_NAM.
Passing AEBODSYS/AEDECOD or AEBODSYS/FMQ_NAM may not give expected results.
\item Passing "Y" to \code{riskyn} argument:
Includes pair-wise risk statistic given by \code{ui_statistics} (Risk Difference or Risk Ratio).
Only the treatments passed in \code{ctrlgrp} and \code{trtgrp} will be considered as the pair.
\code{ui_pctdisp} is automatically set to "TRT" (Risk/percent is calculated for each treatment)
\item Passing "N" to \code{riskyn} argument:
No risk statistic, only incidence counts and percentage are displayed for all treatments.
Arguments \code{ctrlgrp}, \code{trtgrp}, \code{ui_statistics} and \code{ui_alpha} are not required
Only "Count" and "Percent" are permitted for \code{ui_sortvar}.
\code{ui_pctdisp} determines the denominator to calculate \%. It can be "TRT"
(Denominator is Treatment total counts), "VAR" (Variable Total of all treatments/groups)
or "HT" (High Term uses Treatment counts and Low Term uses High Term counts per treatment)
\item \code{ui_sortopt} and \code{ui_sortvar} act together. For example, giving "Ascending" and
"Count" will sort the terms in ascending order of counts in the table.
If \code{ui_sortopt} is "Alphabetical", \code{ui_sortvar} is not considered.
}
}
\examples{
library(cvars)
data(ae_pre)
ae_out <- adae_r001(
  datain = ae_pre,
  population = "Overall",
  AE_Filter = "",
  riskyn = "Y",
  ctrlgrp = "Placebo",
  trtgrp = "Xanomeline Low Dose",
  ui_statistics = "Risk Ratio",
  ui_lt = "AEDECOD",
  ui_ht = "AEBODSYS",
  ui_cutoff = 5
)

ae_out$tout

}
