% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{ADLB}
\format{
Data frame with 74264 features and 46 fields
}
\source{
\url{https://github.com/phuse-org/aesummaries/tree/main/inst/extdata/adlb.xpt},
downloaded 2023-03-17
}
\usage{
adlb
}
\description{
Laboratory Results Chemistry Analysis Dataset.
}
\keyword{datasets}
