% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{ADSL}
\format{
Data frame with 254 features and 58 fields
}
\source{
\url{https://github.com/phuse-org/aesummaries/tree/main/inst/extdata/adsl.xpt},
downloaded 2023-03-17
}
\usage{
adsl
}
\description{
Subject Level Analysis Dataset.
}
\keyword{datasets}
