% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae_pre}
\alias{ae_pre}
\title{ae_pre}
\format{
List of data frames of length \code{3}
}
\usage{
ae_pre
}
\description{
Pre Processed \code{adae} data from \code{ae_pre_processor()}
}
\keyword{datasets}
