% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_pre_processor.R
\name{ae_pre_processor}
\alias{ae_pre_processor}
\title{Pre-processing data for Adverse Event domain specific reports}
\usage{
ae_pre_processor(
  datain,
  ae_filter = "Any Event",
  aeSubset = "AOCCPFL=='Y'",
  aeDenomSubset = "!is.na(ASTDT)",
  aeObsPeriod = "Overall Duration",
  aeObsResidual = 0,
  trtvar = "TRTA",
  trtsort = "TRTAN",
  pop_fil = "SAFFL",
  fmq_data = NA,
  aeEventVar = "AEDECOD",
  aeByVar = "AEBODSYS",
  aeSubGrpVar = NA,
  aeBigN = "Y",
  aeGrpVarMiss = "Y",
  aeTrtTot = "Y",
  aeSubGrpTot = "Y"
)
}
\arguments{
\item{datain}{the ADaM dataset read from the \code{dataread()} utility is taken as input data.}

\item{ae_filter}{The filter UI elements that are applied for the report is taken as input.}

\item{aeSubset}{Subset conditions for analysis of dependent variable.}

\item{aeDenomSubset}{Subset conditions for overall report}

\item{aeObsPeriod}{applied for considering AE event that happened in specific duration.
if user pass "Overall duration" it will filter the \code{STUDYFL == "Y"} records. "Other" will filter
the period which falls under the start and end date.}

\item{aeObsResidual}{If \code{aeObsResidual} sets to "Other", Use this argument to pass a period
(in numeric) to extend the obseravation period.}

\item{trtvar}{Treatment variable name string assigned as TRTVAR.}

\item{trtsort}{Treatment sorting variable - numeric or character}

\item{pop_fil}{Population Filter for data set.}

\item{fmq_data}{FMQ table data set}

\item{aeEventVar}{Event term variable - Lower Level}

\item{aeByVar}{By variable name string assigns to BYVAR1,BYVAR2, etc}

\item{aeSubGrpVar}{Subgroup variable name string assigns to SUBGRPVAR1, SUBGRPVAR2, etc.,}

\item{aeBigN}{Display big N in column header (y/n).}

\item{aeGrpVarMiss}{Include missing values of by group and sub group variables "Y"/"N"}

\item{aeTrtTot}{Display total treatment column in table or plot (y/n).}

\item{aeSubGrpTot}{Display total AE subgroup column in treatment or plot (y/n)}
}
\value{
: a list containing 3 objects
\itemize{
\item dsin - Processed dataframe output for further utilities, with analysis subset
\item dout - Processed dataframe output to be carried to further utilities; no analysis subset
\item trtbign - table or numeric value for big N count, pass to mdisplay
}
}
\description{
Pre-processing data for Adverse Event domain specific reports
}
\examples{
library(cvars)
data(adae)
data(FMQ_Consolidated_List)
data_pre <- ae_pre_processor(
  datain = adae,
  aeSubset = "AOCCPFL=='Y'",
  aeDenomSubset = "!is.na(ASTDT)",
  ae_filter = "Any Event",
  aeObsPeriod = "Overall Duration",
  aeObsResidual = 0,
  trtvar = "TRTA",
  trtsort = "TRTAN",
  pop_fil = "SAFFL",
  fmq_data = FMQ_Consolidated_List,
  aeEventVar = "AEDECOD",
  aeByVar = "AEBODSYS",
  aeSubGrpVar = NA,
  aeBigN = "Y",
  aeGrpVarMiss = "Y",
  aeTrtTot = "Y",
  aeSubGrpTot = "Y"
)

data_pre$dout
}
