% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm}
\alias{cm}
\title{CM}
\format{
Data frame with 7510 rows and 21 variables
}
\source{
\url{https://github.com/phuse-org/aesummaries/tree/main/inst/extdata/cm.xpt},
downloaded 2023-08-24
}
\usage{
cm
}
\description{
Concomitant Medications Dataset
}
\keyword{datasets}
