% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvars-package.R
\docType{package}
\name{cvars-package}
\alias{cvars}
\alias{cvars-package}
\title{cvars: Clinical Visual Analytics for Review and Submission}
\description{
Development of an open-source tool and package to enable generation of identified interactive plots for clinical review and direct inclusion in submission for regulatory agencies. The initial scope is to develop a package to generate interactive forest and volcano plots for adverse event and FDA Medical Queries (FMQs) analysis outputs for inclusion in submissions to the FDA. This work is a collaboration among the American Statistical Association (ASA), PHUSE, and FDA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phuse-org/aesummaries/}
  \item Report bugs at \url{https://github.com/phuse-org/aesummaries/issues/}
}

}
\author{
\strong{Maintainer}: Smriti Anand \email{xxx@xxx.com}

Authors:
\itemize{
  \item Jayashree V
  \item Sukalpo Saha
  \item Vikram S
  \item Jagadish Katam
  \item Arjun Rubalingam
  \item Amol Bandu Gorde
  \item Gomathi S
}

Other contributors:
\itemize{
  \item Dhivya Kanagaraj [contributor]
  \item Neetu Sangari [contributor]
  \item Aishwarya LR [contributor]
  \item Manali Jadhav [contributor]
  \item Prakasam R [contributor]
}

}
