% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_attrib}
\alias{data_attrib}
\title{Variable Labels and Corresponding names from dataframe}
\usage{
data_attrib(datain)
}
\arguments{
\item{datain}{Input dataframe to get variable labels for}
}
\value{
a dataframe with corresponding label for each variable
}
\description{
Variable Labels and Corresponding names from dataframe
}
\examples{
library(cvars)
data("adae")
data_attrib(adae)
}
