% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{data_read}
\alias{data_read}
\title{Read data from selected path or in-built defaults}
\usage{
data_read(ui_data_source, ui_adam_data)
}
\arguments{
\item{ui_data_source}{Source of the input data. Possible Values: "Local", "Default"}

\item{ui_adam_data}{The data set name; when \code{ui_data_source} is "Default".
When \code{ui_data_source} = "Local"; List containing 2 elements, \verb{$datapath} (input file path) and
\verb{$name} (name of the file/dataset)}
}
\value{
a list containing 2 elements:
\itemize{
\item adam - input data read from the user location as data.frame
\item adam_attrib - data.frame with Variable Names and corresponding labels for \code{adam}
}
}
\description{
This function will read data a local path or default; useful for fileInput()
object from Shiny
}
\examples{
library(cvars)
df <- data_read(ui_data_source = "Default", ui_adam_data = "ADSL")
dplyr::slice_head(df$adam$adsl, n = 10)
}
