% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_plot}
\alias{empty_plot}
\title{Empty plot with message}
\usage{
empty_plot(message = "No data available for these values", fontsize = 8)
}
\arguments{
\item{message}{Required message to be displayed within plot area}

\item{fontsize}{Set the font size of the message}
}
\value{
a list containing 2 objects
\itemize{
\item ptly - Interactive empty plot
\item plot - Static empty plot
}
}
\description{
Empty plot with message
}
\examples{
library(cvars)
empty_plot()
}
