% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_analysis.R
\name{event_analysis}
\alias{event_analysis}
\title{To generate event graphs for FDA Medical queries}
\usage{
event_analysis(
  datain = NULL,
  datain_N = NULL,
  hl_var = "FMQ_NAM",
  hl_val = "erythema",
  hl_scope = "Narrow",
  ll_var = "AEDECOD",
  ll_val = "erythema",
  ll_scope = "Narrow",
  summary_by = "Events",
  ref_line = 2
)
}
\arguments{
\item{datain}{input dataset}

\item{datain_N}{input adsl bign dataset with tretament counts based on the population}

\item{hl_var}{MedDRA queries or high level variables valid value : "FMQ_NAM","AEBODSYS"}

\item{hl_val}{Queries/term name captured in the \code{hl_var} variable, One selection e.g.,
erythema}

\item{hl_scope}{scope of the MedDRA queries; valid value : Narrow or Broad}

\item{ll_var}{MedDRA variable; valid value : AEDECOD}

\item{ll_val}{MedDRA value, One selection e.g., erythema}

\item{ll_scope}{scope of the MedDRA queries; valid value : Narrow or Broad}

\item{summary_by}{set to 'Events' or 'patients' based on the requirement to ass the titles and
footnotes}

\item{ref_line}{y axis position of the Horzondal reference line}
}
\value{
generated the normal plot and interactive plotly
}
\description{
To generate event graphs for FDA Medical queries
}
\examples{
library(cvars)
data("event_df")
goutput <- event_analysis(
  datain = event_df$dsin,
  datain_N = event_df$dout,
  hl_var = "FMQ_NAM",
  hl_val = "ABDOMINAL PAIN",
  hl_scope = "Narrow",
  ll_var = "AEDECOD",
  ll_val = "ABDOMINAL DISCOMFORT",
  ll_scope = "Narrow",
  summary_by = "Events",
  ref_line = 2
)
goutput$plot

}
