% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_df}
\alias{event_df}
\title{event_df}
\format{
List containing 2 data frames \code{dsin} and \code{dout} (to be passed in \code{event_analysis}) and
\code{bigN} value
}
\usage{
event_df
}
\description{
Output from \code{risk_stat()} required for \code{event_analysis()}
}
\keyword{datasets}
