% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FMQ_Consolidated_List}
\alias{FMQ_Consolidated_List}
\title{FMQ Consolidated List}
\format{
Data frame with 11712 features and 7 fields
}
\source{
\url{https://github.com/phuse-org/aesummaries/tree/main/inst/extdata/FMQ_Consolidated_List.csv},
downloaded 2023-03-17
}
\usage{
FMQ_Consolidated_List
}
\description{
Consolidated list of FDA Medical Queries.
}
\keyword{datasets}
