% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest Plot of Adverse Events}
\usage{
forest_plot(
  datain,
  AE_Filter,
  review_by,
  summary_by,
  series_color =
    "black,royalblue2,goldenrod,orchid3,brown,pink,aquamarine1,tan4,skyblue1",
  marker_shape = NA,
  statistics,
  xref = 1,
  pvalcut = 0.05,
  trtbign = "Y",
  scale_trans = "identity"
)
}
\arguments{
\item{datain}{Input data from \code{risk_stat()} output, to be used for plotting}

\item{AE_Filter}{Adverse Events filter passed by user in \code{ae_pre_processor()}}

\item{review_by}{Event Variables (Higher and Lower Term) passed by user;
example: \code{c("AEBODSYS","AEDECOD")}}

\item{summary_by}{'Patients' or 'Events' as passed to \code{risk_stat()}
to determine how counts were summed for risk calculation}

\item{series_color}{Comma-separated string of colors for each treatment in
scatterplot or \code{NA} to use in-built colors}

\item{marker_shape}{Comma-separated string of point symbols for each
treatment in scatterplot. \code{NA} to use in-built}

\item{statistics}{Risk statistic from user as passed to \code{risk_stat()}.
Possible values: \code{"Risk Difference"}, \code{"Risk Ratio"}}

\item{xref}{X intercept Value to draw vertical reference line in risk plot}

\item{pvalcut}{p value cutoff to determine and display significant differences
from control treatment}

\item{trtbign}{To display (N=count) next to treatment name in legend or not. Values: \code{"Y"}/\code{"N"}}

\item{scale_trans}{If log transformation needed for X axis of risk plot.
Possible values: \code{"identity"}/\code{"log10"}/\code{"log2"}/\code{"log"}}
}
\value{
: a list containing 7 objects
\itemize{
\item ptly - Interactive plot output for display
\item plot - Static plot output for saving externally
\item drill_plt - Plot data used to create listing
\item rpt_data - Dataset used to create forest plot, for validation purpose
\item n - Number of adverse event terms in plot to determine plot height
\item title - Report title/header
\item footnote - Report footnote
}
}
\description{
\code{forest_plot} takes arguments for displaying Risk and counts data (from \code{risk_stat})
in the form of a forest plot
}
\details{
\itemize{
\item Arguments \code{AE_Filter}, \code{review_by}, \code{summary_by} and \code{statistics} are all used only
in axis
titles, tick labels or hover information
They should be the same values passed previously in the processing \code{ae_pre_processor()}
and analysis \code{risk_stat()} functions
\item Treatments with Risk statistic of a statistically significant p-value (below \code{pvalcut})
are highlighted as 'Significantly Higher' or 'Significantly Lower' in the legend,
relative to the incidence rate of each term
on the control treatment.
\item Default colors for each treatment in scatter plot are preset in \code{series_color}.
Passing NA will change the palette to the package's default color scheme per \code{g_seriescol()}
To override manually, pass list of colors like this: "red,cyan,gold"
\item Shapes of symbols for each treatment in scatterplot are set to take the
package defaults per \code{g_seriessym()}.
To change them manually, pass list of shapes to \code{marker_shape} in format "1,6,25"
or "triangle,square,trianglefilled"
}
}
\examples{
library(cvars)
data(ae_risk)
fp <- forest_plot(
  datain = ae_risk,
  AE_Filter = "Any",
  review_by = c("AEBODSYS", "AEDECOD"),
  summary_by = "Patients",
  statistics = "Risk Ratio"
)

fp$ptly

}
