% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{forestplot}
\alias{forestplot}
\title{Create a widget forest plot table with comparisons}
\usage{
forestplot(
  data = testdata,
  high_level_col = "AEBODSYS",
  low_level_col = "AEDECOD",
  group1_col = "Group1",
  group2_col = "Group2",
  numerator1_col = "n1",
  numerator2_col = "n2",
  denominator1_col = "N1",
  denominator2_col = "N2",
  test_col = "Test",
  result_col = "Res",
  result_upper_col = "CI_Upper",
  result_lower_col = "CI_Lower",
  p_col = "Pvalue",
  width = "100\%",
  height = NULL,
  elementId = NULL,
  comparisonWidth = 200
)
}
\description{
Create a widget forest plot table with comparisons
}
