% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g_seriessym}
\alias{g_seriessym}
\title{Set Shapes to Series Variable to plot}
\usage{
g_seriessym(gdata, series_shapes = NA, SERIESVAR)
}
\arguments{
\item{gdata}{Data being used to create graph}

\item{series_shapes}{Comma-separated shapes corresponding to values in SERIESVAR}

\item{SERIESVAR}{Series Variable to assign colors to. Examples: "TRTVAR","ETHNIC"
Preferably a variable of type factor with distinct levels}
}
\value{
named list of symbols per level of grouping variable
}
\description{
Assigns selected point shapes to levels of a series variable to be used for plotting and legend.
}
\examples{
data(ae_pre)
g_seriessym(ae_pre$dsin, "triangle,cross,circle,square", "TRTVAR")
}
