% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_interval.R
\name{interval_plot}
\alias{interval_plot}
\title{Patient-Level Interval Plot}
\usage{
interval_plot(
  datain,
  startvar,
  endvar,
  yvar,
  seriesvar,
  subjectid,
  series_color = NA,
  xaxislab = "Start and End Study Day",
  yaxislab = ""
)
}
\arguments{
\item{datain}{Input events/concomitant medication dataset}

\item{startvar}{Event/Treatment Start Day Variable name (X axis)}

\item{endvar}{Event/Treatment End Day Variable name (X axis)}

\item{yvar}{Variable containing event terms or medications to be plotted in Y axis}

\item{seriesvar}{Variable to stratify plot by color or NA to have no stratification}

\item{subjectid}{Selected \code{USUBJID} variable value}

\item{series_color}{List of colors to be assigned to levels in \code{seriesvar} or
NA to use package defaults}

\item{xaxislab}{Label for plot X axis}

\item{yaxislab}{Label for plot Y axis}
}
\value{
: a list containing 3 objects
\itemize{
\item ptly - Interactive plot output for display
\item plot - Static plot output for saving externally
\item rpt_data - Dataset used to create forest plot, for validation purpose
}
}
\description{
Patient-Level Interval Plot
}
\examples{
data(adae)
interval_plot(
  datain = adae,
  startvar = "ASTDY",
  endvar = "AENDY",
  yvar = "AEDECOD",
  seriesvar = "AESEV",
  subjectid = "01-701-1302",
  series_color = NA,
  yaxislab = "Reported Term for the Adverse Event"
)$ptly
}
