% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcatstat.R
\name{mcatstat}
\alias{mcatstat}
\title{Summary analysis for categorical variables
\code{mcatstat()} returns a new dataframe containing counts and percentages
of a categorical analysis variable according to grouping and treatment
variables passed in \code{mentry()}}
\usage{
mcatstat(
  datain = NULL,
  d_datain = NULL,
  ui_uniqid = "USUBJID",
  ui_dptvar = NULL,
  ui_pctdisp = "TRT",
  miss_catyn = "N",
  cum_ctyn = "N",
  total_catyn = "N",
  dptvarn = 1
)
}
\arguments{
\item{datain}{Input data from \code{mentry()} output (dsin) to get counts for each
category}

\item{d_datain}{Input data from \code{mentry()} output (dout) to calculate the
denominator for percentage calculation
of the counts from \code{datain}}

\item{ui_uniqid}{Variable to calculate unique counts of.
Likely values: "USUBJID", "SITEID", NA}

\item{ui_dptvar}{Categorical Analysis variable. eg: "SEX","AEDECOD"}

\item{ui_pctdisp}{Method to calculate denominator (for \%) by.
Possible values: "TRT","VAR","COL","SUBGRP","CAT","NONE","NO","DPTVAR"}

\item{miss_catyn}{To include empty/blank values as "Missing" in categories of
\code{ui_dptvar} variable or not. Values: "Y"/"N"}

\item{cum_ctyn}{To return cumulative frequency instead of individual
frequencies for each category. Possible values: "Y"/"N"}

\item{total_catyn}{To return a 'Total' row for the categories of \code{ui_dptvar}
variable or not. Possible values: "Y"/"N"}

\item{dptvarn}{Number to assign as 'DPTVARN', useful for block sorting when
multiple \code{mcatstat()} outputs are created to be combined.}
}
\value{
a data.frame with counts and/or percentages, passed to
\code{mdisplay()} or for further risk statistic calculation
}
\description{
Summary analysis for categorical variables
\code{mcatstat()} returns a new dataframe containing counts and percentages
of a categorical analysis variable according to grouping and treatment
variables passed in \code{mentry()}
}
\details{
\itemize{
\item Objects passed to \code{datain} and \code{d_datain} can be the $dsin and $dout
return elements respectively from either \code{mentry()} or
\code{ae_pre_processor()}
\item \code{ui_uniqid} is the variable name to get unique counts of. If given as
NA, it sums all observations for the given category. If "USUBJID" then it
calculates the number of unique subjects per category.
\item \code{cum_ctyn} as "Y" to get output value as cumulative frequencies instead
of individual frequencies. If "Y", \code{total_catyn} will be reset to "N"
\item pctdisp has possible values for method to get denominator to calculate
percentage:
NONE/NO: No percent calculation
TRT: Treatment total counts acts as denominator
VAR: Variable Total of all treatments/groups acts as denominator
COL: Column wise denominator - percentage within each
Treatment-Subgroup(s) combination
CAT: Row-wise denominator - percentage within each Bygroup(s)-dptvar
combination
SUBGRP: Percentage within each Treatment-By group(s)-Subgroup(s)
combination
DPTVAR: Percentage within each Treatment-By group(s)-Subgroup(s)-dptvar
combination.
}
}
\examples{
data(ae_pre)

ae_cat <- mcatstat(
  datain = ae_pre$dsin,
  d_datain = ae_pre$dout,
  ui_uniqid = "USUBJID",
  ui_dptvar = "AEDECOD",
  ui_pctdisp = "TRT"
)
ae_cat
}
