% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdisplay.R
\name{mdisplay}
\alias{mdisplay}
\title{Display standard tables}
\usage{
mdisplay(
  datain,
  ui_bylabel = NA,
  ui_catlabel = NA,
  ui_statlabel = NA,
  trtbign = NA,
  extra_df = NULL,
  extra_mergeby = "DPTVAL",
  extra_vars = "",
  colformat = "",
  indent = NULL,
  dptlabel = "  "
)
}
\arguments{
\item{datain}{Input data}

\item{ui_bylabel}{Change BYVAR names to label}

\item{ui_catlabel}{Change cat/dptvar names to label}

\item{ui_statlabel}{Change statistics names to label}

\item{trtbign}{Display big N or not}

\item{extra_df}{Additional dataframe merge by variables given in extra_mergeby}

\item{extra_mergeby}{Variables to merge extra_df by, if present}

\item{extra_vars}{Variables additional to standard present in input data to display at last}

\item{colformat}{String (usually indicating format) to add to column (treatment/total) names}

\item{indent}{List of conditions to indent rows, each element increases padding.
eg list("DPTVALN==3","DPTVARN==1")}

\item{dptlabel}{String to become name of the column containing categories (DPTVAL) in output}
}
\value{
a list containing following objects
\itemize{
\item tout - Flextable output to be displayed or downloaded
\item dataf - Final dataframe passed to flextable
}
}
\description{
' mdisplay' is used to generate flextable output of standardised tables
}
\details{
This generic function takes data input from standard Mcatstat/Msumstat utilities.
The resulting flextable is sorted and displayed according to standard variable names,
including variable DPTVARN for sorting blocks and DPTVALN for sorting inside blocks.
}
\examples{
library(cvars)
data("ae_pre")
# mcatstat output will the input for mdisplay
mcat_out <- mcatstat(
  datain = ae_pre$dsin,
  d_datain = ae_pre$dout,
  ui_uniqid = "USUBJID",
  ui_dptvar = "AEDECOD",
  ui_pctdisp = "TRT",
  miss_catyn = "N",
  cum_ctyn = "N",
  total_catyn = "N",
  dptvarn = 1
)

out <- mdisplay(
  datain = mcat_out,
  ui_bylabel = NA,
  ui_catlabel = NA,
  ui_statlabel = NA,
  trtbign = NA,
  extra_df = NULL,
  extra_mergeby = "DPTVAL",
  extra_vars = "",
  colformat = "",
  indent = NULL,
  dptlabel = "  "
)

out$tout

}
