% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentry.R
\name{mentry}
\alias{mentry}
\title{Function to read in and process data with subsets and variables.}
\usage{
mentry(
  datain,
  ui_aSubset,
  ui_dSubset,
  ui_byvar = NA,
  ui_subgrpvar = NA,
  ui_trtvar = NA,
  ui_trtsort = NA,
  ui_trttotalyn = "N",
  ui_sgtotalyn = "N",
  ui_bign = "Y",
  ui_addGrpMiss,
  ui_pop_fil
)
}
\arguments{
\item{datain}{Input dataset - can be any ADaM data}

\item{ui_aSubset}{Subset conditions for analysis of dependent variable}

\item{ui_dSubset}{Subset conditions for overall report}

\item{ui_byvar}{By variable name string assigns to BYVAR1,BYVAR2, etc}

\item{ui_subgrpvar}{Subgroup variable name string assigns to SUBGRPVAR1, SUBGRPVAR2, etc}

\item{ui_trtvar}{Treatment variable name string assigned as TRTVAR}

\item{ui_trtsort}{Treatment sorting variable - numeric or character}

\item{ui_trttotalyn}{Display total treatment column in table or plot (y/n)}

\item{ui_sgtotalyn}{Display total subgroup column in treatment or plot (y/n)}

\item{ui_bign}{Display big N in column header (y/n)}

\item{ui_addGrpMiss}{Add row or column for missing category in grouping variables}

\item{ui_pop_fil}{Population Filter for dataset}
}
\value{
a list containing 3 objects
\itemize{
\item \code{dsin}- processed dataframe output to be carried to further utilities
\item \code{bign}- table or numeric value for big N count, pass to mdisplay
\item \code{dout}- Subset data used for denominator calculation in categorical analysis
}
}
\description{
It takes arguments for TRT,BY and/or SUBGROUP variables, subset conditions and processes input
data. This data flows into categorical and numerical utilities
}
\details{
\enumerate{
\item By Variabel Processing
Identifying byvars if byvar is given, create BYVAR1,2,etc
If byvars also have a corresponding numeric code list variable then create BYVAR1N..
for sorting purpose check for numeric code list variable for the subgroup var,
if not create numeric variable
Convert to character if numeric
Re-code blanks from Bygrp vars as Missing
remove missing category value from grouping variable as per user choice
\item Treatment Variable Processing
Identifying Treatment Variable,treatment sorting variable
removed non treatment values
If treatment sort variable doesn't exist, make it same as TRTVAR
If treat sort is not numeric, convert to numeric
treatment total column
Convert TRTVAR to a factor in same order as TRTSORT
\item Subgroup Variable processing
Identifying Subgroup variable
check for numeric code list variable for the subgroup var, if not create numeric variable
Re-code blanks from By and Subgroup vars as Missing
remove missing category value from grouping variable as per user choice
Subgroup total column for the last subgroup variable
\item Subset Processing
applying population subset selected
applying denominator/over all subset condition passed by the user
applying analysis subset condition passed by the user
\item Calculate big N for mdisplay
big n if treatment exist or subgroup exists
big n if both treatment & Subgroup don't exist
If no big N required then NA
}
}
\examples{
library(cvars)
data(adsl)
datain <- mentry(
  datain = adsl,
  ui_aSubset = "EFFFL=='Y'",
  ui_dSubset = NA,
  ui_byvar = "SEX,AGEGR1",
  ui_trtvar = "TRT01A",
  ui_trtsort = "TRT01AN",
  ui_subgrpvar = "SITEGR1,BMIBLGR1",
  ui_bign = "Y",
  ui_trttotalyn = "N",
  ui_addGrpMiss = "N",
  ui_sgtotalyn = "N",
  ui_pop_fil = "Overall Population"
)

datain$dsin
datain$dout
datain$bign

}
