% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_interval.R
\name{multi_interval}
\alias{multi_interval}
\title{Aligned Interval Plot for Adverse Events and Concomitant Medication}
\usage{
multi_interval(datain, subjectid)
}
\arguments{
\item{datain}{Input list containing an element named \code{AE} (required) and an element named \code{CM}
with the corresponding dataframes}

\item{subjectid}{Participant ID (from the variable \code{USUBJID})}
}
\value{
: a list containing 4 objects
\itemize{
\item ptly - Interactive plot output for display
\item plot - Static plot output for saving externally
\item rpt_data - Dataset used to create plot(s), for validation purpose
\item title - Report title/header, for download purpose
}
}
\description{
Aligned Interval Plot for Adverse Events and Concomitant Medication
}
\examples{
library(cvars)
data(adae)
data(cm)
multi_interval(
  datain = list(AE = adae, CM = cm),
  subjectid = "01-717-1446"
)$ptly
}
