% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Reverse Log transformation of value to pass to scale options}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{Logarithmic base value.}
}
\value{
Transformation object per given base
}
\description{
Reverse Log transformation of value to pass to scale options
}
\examples{
library(cvars)
library(ggplot2)
ggplot(data = mtcars, mapping = aes(x = mpg, y = hp)) +
  geom_point() +
  scale_y_continuous(trans = reverselog_trans(10))
}
