% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_stat.R
\name{risk_stat}
\alias{risk_stat}
\title{Function to calculate risk statistics for treatment pairs from AE pre-processed data}
\usage{
risk_stat(
  datain,
  d_datain,
  summary_by,
  eventVar,
  ctrlgrp,
  trtgrp,
  statistics,
  alpha,
  cutoff,
  sort_opt,
  sort_var
)
}
\arguments{
\item{datain}{Input dataset after pre_processing of AE data with analysis subset applied}

\item{d_datain}{Input dataset after pre_processing of AE data for denominator}

\item{summary_by}{Measure to construct the summary by. Values: 'Patients' or 'Events'}

\item{eventVar}{Event Variable to review by. Example: "AEDECOD", "AEBODSYS"}

\item{ctrlgrp}{Control treatment value}

\item{trtgrp}{Treatment(s) to create pairs. Only 1 value for Volcano/table,
can be multiple for Forest}

\item{statistics}{Statistic to be calculated. Values: 'Risk Ratio' or 'Risk Difference'}

\item{alpha}{Alpha value to determine confidence interval for risk calculation. Default: 0.05}

\item{cutoff}{Incidence Cutoff Value - consider only terms with incidence percentage > ui_cutoff}

\item{sort_opt}{How to sort terms, only for table/forest plot.
Values: 'Ascending','Descending','Alphabetical'}

\item{sort_var}{Metric to sort by. Values: 'Count','Percent','RiskValue'}
}
\value{
a dataset containing risk statistic calculations for given treatment pair(s)
}
\description{
Function to calculate risk statistics for treatment pairs from AE pre-processed data
}
\examples{
library(cvars)

risk <- risk_stat(
  datain = ae_pre$dsin,
  d_datain = ae_pre$dout,
  summary_by = "Patients",
  eventVar = "AEDECOD",
  ctrlgrp = "Placebo",
  trtgrp = "Xanomeline High Dose",
  statistics = "Risk Ratio",
  alpha = 0.05,
  cutoff = 2,
  sort_opt = "Ascending",
  sort_var = "Count"
)

risk[25:35, ]
}
