% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskdiff_wald.R
\name{riskdiff_wald}
\alias{riskdiff_wald}
\title{Calculate Risk difference}
\usage{
riskdiff_wald(
  x,
  y = NULL,
  conf.level = 0.95,
  rev = c("neither", "rows", "columns", "both"),
  correction = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{input data
input data can be one of the following: r x 2 table, vector of numbers from a
contigency table (will be transformed into r x 2 table in row-wise order),
or single factor or character vector that will be combined with y into a table.}

\item{y}{single factor or character vector that will be combined with x into a table
(default is NULL)}

\item{conf.level}{confidence level (default is 0.95)}

\item{rev}{reverse order of "rows", "colums", "both", or "neither" (default)}

\item{correction}{Yate's continuity correction}

\item{verbose}{To return more detailed results}
}
\value{
a  list containg a data,measure,p.value,correction
}
\description{
Function to calculate risk difference by unconditional maximum likelihood estimation (Wald)
for any given treatment pairs.
}
\examples{
riskdiff_wald(
  x = matrix(c(178, 79, 1411, 1486), 2, 2),
  conf.level = 0.95,
  rev = c("neither", "rows", "columns", "both"),
  correction = FALSE,
  verbose = FALSE
)
}
