% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_file.R
\name{save_file}
\alias{save_file}
\title{To save the reports generated in the application}
\usage{
save_file(save_object, file_format, report_type, report_name, file)
}
\arguments{
\item{save_object}{list of list containing, the Plot and Plotly, or flex table of the report
generated along with Tile and footnote}

\item{file_format}{the format in with file should be saved , possible values pdf, html, doc,
ppt, interactive html
(\verb{for Figures}) \emph{Permitted Values}: \code{"pdf"}, \code{"html"}, \code{"docx"}, \code{"pptx"}, \code{"interactive"}
(\verb{for Tables}) \emph{Permitted Values}: \code{"pdf"}, \code{"html"}, \code{"docx"}}

\item{report_type}{To identify whether the report type is table , figure is to be downloaded
\emph{Permitted Values}: \code{"table"}, \code{"figure"}}

\item{report_name}{Name of the report that is created formt he report meta list}

\item{file}{Name in which the report output file need to be saved}
}
\value{
auto save on click
}
\description{
To save the reports generated in the application
}
\examples{
\dontrun{
save_file (
save_object = <shiny reactive graph/table object>,
file_format = 'pdf',
report_name = "Forest Plot",
report_type = 'Figure',
file = <reportname, timestamp, format>
)
}
}
