% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_start}
\alias{var_start}
\title{Find column names with starting pattern}
\usage{
var_start(df, pattern)
}
\arguments{
\item{df}{Dataframe input from which to get names}

\item{pattern}{String. If not ends with N,only matches start. If ends with N, matches start and
checks if column name ends in "N"}
}
\value{
Vector of column names starting with given pattern
}
\description{
Searches column names of given dataframe to identify which start with the given pattern.
If the pattern given ends with 'N', it looks for variables that also end with N.
If the pattern does not end with "N", it excludes variables that end with "N" while returning
those that start with the pattern. eg. pattern="BYVAR" may return c("BYVAR1","BYVAR2")
Whereas "BYVARN" may return c("BYVAR1N","BYVAR2N"), if they exist.
}
\examples{
data(ae_pre)
var_start(ae_pre$dsin, "BYVARN")
var_start(ae_pre$dsin, "BYVAR")
}
