% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Function to generate Volcano Plot of Adverse Events data from Risk values of treatment pairs}
\usage{
volcano_plot(
  datain,
  AE_Filter,
  statistics_data,
  statistics,
  treatment1,
  treatment2,
  X_ref,
  summary_by,
  pvalue_label,
  treatment1_label,
  treatment2_label,
  pvalcut
)
}
\arguments{
\item{datain}{(\verb{list of data.frame})\cr AE dataset after pre-processing for filters
using \code{ae_pre_processor()}.}

\item{AE_Filter}{(\code{character})\cr Adverse Event filter selected, used for footnote.}

\item{statistics_data}{(\code{data.frame})\cr Input dataset containing statistics for treatment
pairs from AE data, generated by \code{risk_stat()}.}

\item{statistics}{(\code{character})\cr Statistic which was calculated in statistics_data.
Expected Values: \verb{Risk Difference}/\verb{Risk Ratio}.}

\item{treatment1}{(\code{character})\cr Control treatment.}

\item{treatment2}{(\code{character})\cr Exposure/comparison treatment.}

\item{X_ref}{(\code{numeric})\cr X intercept value to draw vertical reference lines in plot.}

\item{summary_by}{(\code{character})\cr Calculated summaries by; used for footnote.
Expected Values: \code{Patients}/\code{Events}.}

\item{pvalue_label}{(\code{character})\cr P-value transformation. Expected Values: \code{-log10}, \code{None}}

\item{treatment1_label}{(\code{character})\cr Label for treatment1 in plot title}

\item{treatment2_label}{(\code{character})\cr Label for treatment2 in plot title}

\item{pvalcut}{(\code{numeric})\cr p-value cut-off, filter values with adjusted \code{p} less than
\code{pvalcut}.}
}
\value{
a list containing 5 elements
\itemize{
\item ptly - Interactive plotly output to be displayed
\item plot - Static plot output
\item rpt_data - Data used to create report for download/QC
\item title - String to use as report title
\item footnote - String to use as report footnote
}
}
\description{
Function to generate Volcano Plot of Adverse Events data from Risk values of treatment pairs
}
\examples{

library(cvars)

vout <- volcano_plot(
  datain = ae_pre,
  AE_Filter = "Safety",
  statistics_data = ae_risk,
  statistics = "Risk Difference",
  treatment1 = "Placebo",
  treatment2 = "Xanomeline Low Dose",
  X_ref = 1,
  summary_by = "Patients",
  pvalue_label = "None",
  treatment1_label = "Control",
  treatment2_label = "Exposure",
  pvalcut = 0.05
)

vout$ptly

}
