% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvars-package.R
\docType{package}
\name{cvars-package}
\alias{cvars}
\alias{cvars-package}
\title{cvars: Clinical Visual Analytics for Review and Submission}
\description{
The purpose of this Application is to visualize the AE (Adverse Event) dataset through Interactive R shiny Graphs (Volcano, Forest and Event Analysis Plot). In this Application, we can generate the plots which will be according to the filters that we select. Since we have many table of tables to visualize certain statistics and measures for AE, we can simply use this app to display plots for all the combinations of treatments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phuse-org/aesummaries/}
  \item Report bugs at \url{https://github.com/phuse-org/aesummaries/issues/}
}

}
\author{
\strong{Maintainer}: Smriti Anand \email{SMRITI.ANAND@pfizer.com}

Authors:
\itemize{
  \item Jayashree V
  \item Sukalpo Saha
  \item Vikram S
  \item Jagadish Katam
  \item Arjun Rubalingam
  \item Amol Bandu Gorde
  \item Gomathi S
}

Other contributors:
\itemize{
  \item Dhivya Kanagaraj [contributor]
  \item Manali Jadhav [contributor]
  \item Prakasam R [contributor]
}

}
